Groq NLP added
added Groq AI. so now user do not have to wait for the agent to ask multiple questions.
User can say a sentence that he wants to book a table for 4 people for tomorrow and the time and cooking preference and any special request or mention seating area. The agent will  interpret from the sentence and ask the user if he wants to confirm the booking.